+++ frontend/src/App.tsx
@@ -1,9 +1,8 @@
 import "./App.css";
 
-import React from "react";
-import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-
-import Footer from "./containers/footer/Footer";
+import React, { useState, useEffect } from "react";
+import { BrowserRouter as Router, Route, Switch, Redirect } from "react-router-dom";
+// import Footer from "./containers/footer/Footer";
 import Header from "./containers/header/Header";
 import CreatePost from "./pages/create_post/CreatePost";
 import Home from "./pages/home";
@@ -16,6 +15,12 @@ import { ViewProfile } from "./pages/view_profile/ViewProfile";
 import { LogOut } from "./pages/login/Logout";
 import { Experimental } from "./pages/experimental/Experimental";
 import { LoginGitHub } from "./pages/login/LoginGitHub";
+import { AppDispatch } from "./store";
+import { useDispatch } from "react-redux";
+import { fetchTrendingPosts } from "./reducers/postsSlice";
+import { unwrapResult } from "@reduxjs/toolkit";
+import { Loading } from "./pages/common/Loading";
+import ErrorPage from "./pages/common/ErrorPage";
 
 export type ViewProfileParams = {
   username: string;
@@ -26,13 +31,36 @@ export type PostDetailParams = {
 };
 
 export default function App() {
+  const [loading, setLoading] = useState(true);
+  const [error, setError] = useState(null);
+
+  const dispatch = useDispatch<AppDispatch>();
+  useEffect(() => {
+    setLoading(true);
+    dispatch(fetchTrendingPosts())
+      .then(unwrapResult).then( //must set dispatch to any to use .then
+        () => {
+          setLoading(false)
+        }
+      ).catch(error => {
+        console.log(error);
+        setError(error);
+        setLoading(false);
+      });
+  }, []);
+
+  if (loading) {
+    return <Loading />
+  }
+
+  if (error) {
+    return <ErrorPage error={error} />
+  }
+
   return (
     <Router>
       <Header></Header>
       <Switch>
-        <Route exact path="/">
-          <Home />
-        </Route>
         <Route path="/user/:username">
           <ViewProfile />
         </Route>
@@ -65,6 +93,12 @@ export default function App() {
         </Route>
         <Route path="/test">
           <Experimental />
+        </Route>
+        <Route path="/home">
+          <Home />
+        </Route>
+        <Route exact path="/">
+          <Redirect to="/home"></Redirect>
         </Route>
       </Switch>
       {/* <Footer></Footer> */}
